Attempt #5
This wasn't quite enough to pass Challenge #5. I saw results in the high
90s but never quite reached 100.

The biggest difference is that I'm now tracking whether the up or down
button was pressed, and setting the up/down indicators on each elevator
to affect which passengers board when stopping on a floor.

Occasionally an elevator would stop on a floor and not start moving
again. I'm assuming this happens when there's nothing tracked in the
waitingFloors queue at the time the "idle" event is processed. There's
nothing to tell the elevator to try again.

I think the best move forward is to mark the elevator as idle, then in
the update function check if we have any idle elevators and assign
destination queues there.